requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
dateparser==0.6.0
future==0.16.0
httplib2==0.10.3
humanize==0.5.1
lxml==3.8.0
mailmanclient==3.1
maya==0.3.2
pendulum==1.2.4
python-dateutil==2.6.1
python-telegram-bot==7.0.0
pytz==2017.2
pytzdata==2017.2
regex==2017.7.28
ruamel.yaml==0.15.22
six==1.10.0
tzlocal==1.4
